SegFormer/requirements.txt
timm
einops
tabulate
tensorboard
opencv-python
matplotlib==3.8.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.26.3
pyarrow==15.0.0
pillow==10.2.0
tqdm==4.66.2

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.0
torchvision==0.17.0
